+++ hardware/vivado/ip/fm_receiver_hls/drivers/fm_receiver_hls_v0_1/data/fm_receiver_hls.tcl
@@ -6,17 +6,17 @@
 # ==============================================================
 
 proc generate {drv_handle} {
-    xdefine_include_file $drv_handle "xparameters.h" "XFm_receiver_top" \
+    xdefine_include_file $drv_handle "xparameters.h" "XFm_receiver_hls" \
         "NUM_INSTANCES" \
         "DEVICE_ID" \
         "C_S_AXI_CONFIG_BASEADDR" \
         "C_S_AXI_CONFIG_HIGHADDR"
 
-    xdefine_config_file $drv_handle "xfm_receiver_top_g.c" "XFm_receiver_top" \
+    xdefine_config_file $drv_handle "xfm_receiver_hls_g.c" "XFm_receiver_hls" \
         "DEVICE_ID" \
         "C_S_AXI_CONFIG_BASEADDR"
 
-    xdefine_canonical_xpars $drv_handle "xparameters.h" "XFm_receiver_top" \
+    xdefine_canonical_xpars $drv_handle "xparameters.h" "XFm_receiver_hls" \
         "DEVICE_ID" \
         "C_S_AXI_CONFIG_BASEADDR" \
         "C_S_AXI_CONFIG_HIGHADDR"
