+++ hardware/vivado/ip/fm_receiver_hls/drivers/fm_receiver_hls_v0_1/src/xfm_receiver_hls.c
@@ -1,0 +1,40 @@
+// ==============================================================
+// File generated by Vivado(TM) HLS - High-Level Synthesis from C, C++ and SystemC
+// Version: 2018.2
+// Copyright (C) 1986-2018 Xilinx, Inc. All Rights Reserved.
+// 
+// ==============================================================
+
+/***************************** Include Files *********************************/
+#include "xfm_receiver_hls.h"
+
+/************************** Function Implementation *************************/
+#ifndef __linux__
+int XFm_receiver_hls_CfgInitialize(XFm_receiver_hls *InstancePtr, XFm_receiver_hls_Config *ConfigPtr) {
+    Xil_AssertNonvoid(InstancePtr != NULL);
+    Xil_AssertNonvoid(ConfigPtr != NULL);
+
+    InstancePtr->Config_BaseAddress = ConfigPtr->Config_BaseAddress;
+    InstancePtr->IsReady = XIL_COMPONENT_IS_READY;
+
+    return XST_SUCCESS;
+}
+#endif
+
+void XFm_receiver_hls_Set_led_ctrl(XFm_receiver_hls *InstancePtr, u32 Data) {
+    Xil_AssertVoid(InstancePtr != NULL);
+    Xil_AssertVoid(InstancePtr->IsReady == XIL_COMPONENT_IS_READY);
+
+    XFm_receiver_hls_WriteReg(InstancePtr->Config_BaseAddress, XFM_RECEIVER_HLS_CONFIG_ADDR_LED_CTRL_DATA, Data);
+}
+
+u32 XFm_receiver_hls_Get_led_ctrl(XFm_receiver_hls *InstancePtr) {
+    u32 Data;
+
+    Xil_AssertNonvoid(InstancePtr != NULL);
+    Xil_AssertNonvoid(InstancePtr->IsReady == XIL_COMPONENT_IS_READY);
+
+    Data = XFm_receiver_hls_ReadReg(InstancePtr->Config_BaseAddress, XFM_RECEIVER_HLS_CONFIG_ADDR_LED_CTRL_DATA);
+    return Data;
+}
+
