+++ hardware/vivado/ip/fm_receiver_hls/drivers/fm_receiver_hls_v0_1/src/xfm_receiver_hls_hw.h
@@ -1,0 +1,21 @@
+// ==============================================================
+// File generated by Vivado(TM) HLS - High-Level Synthesis from C, C++ and SystemC
+// Version: 2018.2
+// Copyright (C) 1986-2018 Xilinx, Inc. All Rights Reserved.
+// 
+// ==============================================================
+
+// CONFIG
+// 0x00 : reserved
+// 0x04 : reserved
+// 0x08 : reserved
+// 0x0c : reserved
+// 0x10 : Data signal of led_ctrl
+//        bit 7~0 - led_ctrl[7:0] (Read/Write)
+//        others  - reserved
+// 0x14 : reserved
+// (SC = Self Clear, COR = Clear on Read, TOW = Toggle on Write, COH = Clear on Handshake)
+
+#define XFM_RECEIVER_HLS_CONFIG_ADDR_LED_CTRL_DATA 0x10
+#define XFM_RECEIVER_HLS_CONFIG_BITS_LED_CTRL_DATA 8
+
