+++ hardware/vivado/ip/fm_receiver_hls/drivers/fm_receiver_hls_v0_1/src/xfm_receiver_hls_sinit.c
@@ -1,0 +1,46 @@
+// ==============================================================
+// File generated by Vivado(TM) HLS - High-Level Synthesis from C, C++ and SystemC
+// Version: 2018.2
+// Copyright (C) 1986-2018 Xilinx, Inc. All Rights Reserved.
+// 
+// ==============================================================
+
+#ifndef __linux__
+
+#include "xstatus.h"
+#include "xparameters.h"
+#include "xfm_receiver_hls.h"
+
+extern XFm_receiver_hls_Config XFm_receiver_hls_ConfigTable[];
+
+XFm_receiver_hls_Config *XFm_receiver_hls_LookupConfig(u16 DeviceId) {
+	XFm_receiver_hls_Config *ConfigPtr = NULL;
+
+	int Index;
+
+	for (Index = 0; Index < XPAR_XFM_RECEIVER_HLS_NUM_INSTANCES; Index++) {
+		if (XFm_receiver_hls_ConfigTable[Index].DeviceId == DeviceId) {
+			ConfigPtr = &XFm_receiver_hls_ConfigTable[Index];
+			break;
+		}
+	}
+
+	return ConfigPtr;
+}
+
+int XFm_receiver_hls_Initialize(XFm_receiver_hls *InstancePtr, u16 DeviceId) {
+	XFm_receiver_hls_Config *ConfigPtr;
+
+	Xil_AssertNonvoid(InstancePtr != NULL);
+
+	ConfigPtr = XFm_receiver_hls_LookupConfig(DeviceId);
+	if (ConfigPtr == NULL) {
+		InstancePtr->IsReady = 0;
+		return (XST_DEVICE_NOT_FOUND);
+	}
+
+	return XFm_receiver_hls_CfgInitialize(InstancePtr, ConfigPtr);
+}
+
+#endif
+
