--- hardware/vivado/ip/fm_receiver_hls/drivers/fm_receiver_top_v0_1/src/xfm_receiver_top.h
@@ -1,92 +1,0 @@
-// ==============================================================
-// File generated by Vivado(TM) HLS - High-Level Synthesis from C, C++ and
-// SystemC Version: 2018.2 Copyright (C) 1986-2018 Xilinx, Inc. All Rights
-// Reserved.
-//
-// ==============================================================
-
-#ifndef XFM_RECEIVER_TOP_H
-#define XFM_RECEIVER_TOP_H
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-#undef __linux__
-/***************************** Include Files *********************************/
-#ifndef __linux__
-#include "xil_assert.h"
-#include "xil_io.h"
-#include "xil_types.h"
-#include "xstatus.h"
-#else
-#include <assert.h>
-#include <dirent.h>
-#include <fcntl.h>
-#include <stddef.h>
-#include <stdint.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#include <sys/mman.h>
-#include <unistd.h>
-#endif
-#include "xfm_receiver_top_hw.h"
-
-/**************************** Type Definitions ******************************/
-#ifdef __linux__
-typedef uint8_t u8;
-typedef uint16_t u16;
-typedef uint32_t u32;
-#else
-typedef struct {
-  u16 DeviceId;
-  u32 Config_BaseAddress;
-} XFm_receiver_top_Config;
-#endif
-
-typedef struct {
-  u32 Config_BaseAddress;
-  u32 IsReady;
-} XFm_receiver_top;
-
-/***************** Macros (Inline Functions) Definitions *********************/
-#ifndef __linux__
-#define XFm_receiver_top_WriteReg(BaseAddress, RegOffset, Data) \
-  Xil_Out32((BaseAddress) + (RegOffset), (u32)(Data))
-#define XFm_receiver_top_ReadReg(BaseAddress, RegOffset) \
-  Xil_In32((BaseAddress) + (RegOffset))
-#else
-#define XFm_receiver_top_WriteReg(BaseAddress, RegOffset, Data) \
-  *(volatile u32 *)((BaseAddress) + (RegOffset)) = (u32)(Data)
-#define XFm_receiver_top_ReadReg(BaseAddress, RegOffset) \
-  *(volatile u32 *)((BaseAddress) + (RegOffset))
-
-#define Xil_AssertVoid(expr)    assert(expr)
-#define Xil_AssertNonvoid(expr) assert(expr)
-
-#define XST_SUCCESS            0
-#define XST_DEVICE_NOT_FOUND   2
-#define XST_OPEN_DEVICE_FAILED 3
-#define XIL_COMPONENT_IS_READY 1
-#endif
-
-/************************** Function Prototypes *****************************/
-#ifndef __linux__
-int XFm_receiver_top_Initialize(XFm_receiver_top *InstancePtr, u16 DeviceId);
-XFm_receiver_top_Config *XFm_receiver_top_LookupConfig(u16 DeviceId);
-int XFm_receiver_top_CfgInitialize(XFm_receiver_top *InstancePtr,
-                                   XFm_receiver_top_Config *ConfigPtr);
-#else
-int XFm_receiver_top_Initialize(XFm_receiver_top *InstancePtr,
-                                const char *InstanceName);
-int XFm_receiver_top_Release(XFm_receiver_top *InstancePtr);
-#endif
-
-void XFm_receiver_top_Set_led_ctrl(XFm_receiver_top *InstancePtr, u32 Data);
-u32 XFm_receiver_top_Get_led_ctrl(XFm_receiver_top *InstancePtr);
-
-#ifdef __cplusplus
-}
-#endif
-
-#endif
